Move form-field and user-name-alias to components/
