+++ packages/angular/projects/ui-angular/src/lib/primitives/index.ts
@@ -1,10 +1,8 @@
-export { AmplifyFormFieldComponent } from './form-field/form-field.component';
 export { AmplifyErrorComponent } from './error/error.component';
 export { AmplifyButtonComponent } from './button/button.component';
 export { AmplifyPasswordFieldComponent } from './password-field/password-field.component';
 export { AmplifySelectComponent } from './select/select.component';
 export { AmplifyTextFieldComponent } from './text-field/text-field.component';
-export { AmplifyUserNameAliasComponent } from './user-name-alias/user-name-alias.component';
 export { PhoneNumberFieldComponent } from './phone-number-field/phone-number-field.component';
 export { TabItemComponent } from './tab-item/tab-item.component';
 export { TabsComponent } from './tabs/tabs.component';
